Taking an object oriented programming approach.
As of this weekend, each crossing is an individual object that has a
hashmap of strand numbers and colors as well as functions to edit it. I
think this extra organizing will make the makeValid function easier to
code. Right now the main problem is in getting the "makeValid" function
to work at all. The current idea is to have the makeValid function
return a hashmap of changed strand numbers and the goal color so this
can be passed to the other crossings. I started with trying to make a
crossing valid when it only has one unassigned strand color, but
something is missing for now. This will probably be solved within a
week.